Updates the calculateInventoryWeight function to use parseFloat so that we can handle decimal item weights. This is needed for bundles of very light items, like arrows and crossbow bolts.

+++ src/apps/handlers/calculateInventoryWeight.js
@@ -3,7 +3,7 @@ export default function calculateInventoryWeight(actorData) {
 
   const totalItemWeight = equippedItems.reduce((acc, curr) => {
     const { quantity } = curr.system;
-    const itemWeight = parseInt(curr.system.weight, 10) ?? 0;
+    const itemWeight = parseFloat(curr.system.weight, 10) ?? 0;
 
     return acc + (quantity ? itemWeight * quantity : itemWeight);
   }, 0);
